cybcer-soc/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
rich>=13.0.0